Improve abbreviations

+++ config/fish/conf.d/abbreviations.fish
@@ -28,36 +28,42 @@ abbr cd.. 'cd ..'
 
 abbr rmf 'rm -rf'
 
-abbr gut 'git'
-abbr gcl "git clone --recurse-submodules"
-abbr gcma "git cma"
-abbr gcmap "git cmap"
-abbr gd "git diff"
-abbr ggc "git gc"
-abbr gib "git init --bare"
-abbr gl "git lg"
-abbr glc "git diff @~..@"
-abbr gm "git checkout main"
-abbr gp "git pull --progress --no-rebase"
-abbr gpn "git pull --no-ff"
-abbr gpo "git push origin"
-abbr gr "git remote -v"
-abbr gst "git status -sb"
-abbr gsu "git submodule update --recursive --remote"
+if command -sq git
+    abbr gut 'git'
+    abbr gcl "git clone --recurse-submodules"
+    abbr gcma "git cma"
+    abbr gcmap "git cmap"
+    abbr gd "git diff"
+    abbr ggc "git gc"
+    abbr gib "git init --bare"
+    abbr gl "git lg"
+    abbr glc "git diff @~..@"
+    abbr gm "git checkout main"
+    abbr gp "git pull --progress --no-rebase"
+    abbr gpn "git pull --no-ff"
+    abbr gpo "git push origin"
+    abbr gr "git remote -v"
+    abbr gst "git status -sb"
+    abbr gsu "git submodule update --recursive --remote"
+end
 
-abbr docker-compose "docker compose"
-abbr dssh "docker exec -it"
-abbr dcu "docker compose up -d"
-abbr dcuf "docker compose up -d && docker compose logs -f"
-abbr dcd "docker compose down"
-abbr dcp "docker compose pull"
-abbr dcl "docker compose logs -f"
-abbr dps "docker ps --format \"table {{ .Names }}\" -a"
-abbr dsp "docker system prune -a"
-abbr dil "docker image ls"
-abbr dsd "docker system df"
+if command -sq docker
+    abbr docker-compose "docker compose"
+    abbr dssh "docker exec -it"
+    abbr dcu "docker compose up -d"
+    abbr dcuf "docker compose up -d && docker compose logs -f"
+    abbr dcd "docker compose down"
+    abbr dcp "docker compose pull"
+    abbr dcl "docker compose logs -f"
+    abbr dps "docker ps --format \"table {{ .Names }}\" -a"
+    abbr dsp "docker system prune -a"
+    abbr dil "docker image ls"
+    abbr dsd "docker system df"
+end
 
-abbr composer 'composer --ansi'
+if command -sq composer
+    abbr composer 'composer --ansi'
+end
 
 abbr more 'less'
 abbr vom 'vim'
@@ -72,5 +78,7 @@ abbr iptuseage 'sudo iptables -L -nvx | grep -v " 0 DROP"'
 abbr firewall iptlist
 abbr ipt 'sudo /sbin/iptables'
 
-abbr journalctl 'sudo journalctl'
-abbr systemctl 'sudo systemctl'
+if not fish_is_root_user
+    abbr journalctl 'sudo journalctl'
+    abbr systemctl 'sudo systemctl'
+end
